added annos on sections for organic split
added enacted="true" flag to all enacted titles

added enacted="false" to all sections that are too dificult
to move to their organic sections, and that we will therefore
leave in their current location in the code. the enacted="false"
will let us know that even though they are in the code, they
are not actually enacted.

added manual="true" to sections whose organic annotation is
just too "weird" to deal with with the parser. We will go
through and manually handle these organics.